% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNumbers.R
\name{add}
\alias{add}
\title{Add two numbers together}
\usage{
add(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{another number}
}
\value{
the sum of \code{x} \code{y}
}
\description{
Add two numbers together
}
\note{
this example is taken from the roxygen documnetation here: https://cran.r-project.org/web/packages/roxygen2/vignettes/roxygen2.html
}
\examples{
 add(1,1)
 add(10,1)

}
