% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two_numbers.R
\name{add_two_numbers}
\alias{add_two_numbers}
\title{Add Two Numbers}
\usage{
add_two_numbers(a, b)
}
\arguments{
\item{a}{(numeric) A positive number}

\item{b}{(numeric) A positive number}
}
\value{
The total of the two number supplied
}
\description{
The purpose of this function is to take two numbers and add them
\code{add_two_numbers} is an awesome function!
}
\examples{
add_two_numbers(a = 1, b = 1)
}
