% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/does_Philip_ever_panic.R
\name{does_Philip_ever_panic}
\alias{does_Philip_ever_panic}
\title{Whether or not Philip ever panics}
\usage{
does_Philip_ever_panic(x)
}
\arguments{
\item{x}{'yes'/'no'}
}
\value{
a string depending on whether you choose 'yes' or 'no'
}
\description{
This function returns a funny answer depending on whether
you choose 'yes' or 'no' to the famous question:
Does Philip ever panic?
}
\examples{
does_Philip_ever_panic('yes')
does_Philip_ever_panic('no')

}
