% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/does_Tamil_ever_panic.R
\name{does_Tamil_ever_panic}
\alias{does_Tamil_ever_panic}
\title{Whether Tamil ever panics or not}
\usage{
does_Tamil_ever_panic(x)
}
\arguments{
\item{x}{'yes'/'no'}
}
\value{
a string depending on whether you choose 'yes' or 'no'
}
\description{
This function returns a funny answer depending on whether
you choose 'yes' or 'no to the famous question:
Does Tamil ever panic?
}
\examples{
does_Tamil_ever_panic('yes')
does_Tamil_ever_panic('no')

}
