% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_function.R
\name{generic_function}
\alias{generic_function}
\title{Generic Function}
\usage{
generic_function(input_number1, input_number2)
}
\arguments{
\item{input_number1}{A number of your choosing}

\item{input_number2}{Another number of your choosing (can be the same or different)}
}
\value{
A number (about double) with some random noise
}
\description{
A very boring function, but really just here for example.
}
\examples{
noisy_variable <- generic_function(4,3)

}
