% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(first = "Laura", last = "OWI")
}
\arguments{
\item{first}{character The first name}

\item{last}{character The last name}
}
\description{
This is our hello world function!
}
\examples{
hello()
hello("Laura")
hello("Laura","OWI")
hello(last="OWI", first="Laura")

}
