% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_Alex_the_best_CEO_and_mentor.R
\name{is_Alex_the_best_CEO_and_mentor}
\alias{is_Alex_the_best_CEO_and_mentor}
\title{Whether or not Alex is the best CEO and mentor}
\usage{
is_Alex_the_best_CEO_and_mentor(x)
}
\arguments{
\item{x}{'yes'/'no'}
}
\value{
a string depending on whether you choose 'yes' or 'no'
}
\description{
This function returns a funny answer depending on whether
you choose 'yes' or 'no' to the famous question:
is Alex the best CEO and mentor?
}
\examples{
is_Alex_the_best_CEO_and_mentor('yes')
is_Alex_the_best_CEO_and_mentor('no')

}
