% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_Anand_the_best_CEO_and_mentor.R
\name{is_Anand_the_best_CEO_and_mentor}
\alias{is_Anand_the_best_CEO_and_mentor}
\title{Whether Anand Sir is the best CEO and mentor or not}
\usage{
is_Anand_the_best_CEO_and_mentor(x)
}
\arguments{
\item{x}{'yes'/'no'}
}
\value{
a string depending on whether you choose 'yes' or 'no'
}
\description{
This function returns a funny answer depending on whether
you choose 'yes' or 'no to the famous question:
is Anand Sir the best CEO and mentor?
}
\examples{
is_Anand_the_best_CEO_and_mentor('yes')
is_Anand_the_best_CEO_and_mentor('no')

}
