% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_Sailaja_frightening.R
\name{is_Sailaja_frightening}
\alias{is_Sailaja_frightening}
\title{Is Sailaja frightening or not}
\usage{
is_Sailaja_frightening(x)
}
\arguments{
\item{x}{'yes'/'no'}
}
\value{
a string depending on whether you choose 'yes' or 'no'
}
\description{
This function returns a funny answer depending on whether
you choose 'yes' or 'no to the famous question:
Is Sailaja frightening?
}
\examples{
is_Sailaja_frightening('yes')
is_Sailaja_frightening('no')

}
