% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_Satan_frightening.R
\name{is_Satan_frightening}
\alias{is_Satan_frightening}
\title{Is Satan frightening or not}
\usage{
is_Satan_frightening(x)
}
\arguments{
\item{x}{'yes'/'no'}
}
\value{
a string depending on whether you choose 'yes' or 'no'
}
\description{
This function returns a funny answer depending on whether
you choose 'yes' or 'no' to the famous question:
Is Satan frightening?
}
\examples{
is_Satan_frightening('yes')
is_Satan_frightening('no')

}
