% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{lookup_thingie}
\alias{lookup_thingie}
\title{Example lookup_thingie}
\description{
Example data representing lookup_thingie
}
\examples{
head(lookup_thingie)
}
\keyword{data}
\keyword{quality}
\keyword{water}
