% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange_numbers.R
\name{rearrange_numbers}
\alias{rearrange_numbers}
\title{rearrange_numbers}
\usage{
rearrange_numbers(numberVec)
}
\arguments{
\item{numberVec}{Vector of numbers}
}
\value{
New vector of same length as input.
}
\description{
rearrange_numbers
}
