% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangeNumbers.R
\name{rearrangeNumbers}
\alias{rearrangeNumbers}
\title{rearrangeNumbers}
\usage{
rearrangeNumbers(numberVec)
}
\arguments{
\item{numberVec}{Vector of numbers}
}
\value{
New vector of same length as input.
}
\description{
rearrangeNumbers
}
