% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_dv.R
\name{summary_dv}
\alias{summary_dv}
\title{Summary DD}
\usage{
summary_dv(site)
}
\arguments{
\item{site}{character site number}
}
\description{
calculate max and min of year
}
\examples{
sum_table <- summary_dv("05398000")



}
