% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(first, last = "userbot")
}
\arguments{
\item{last}{}
}
\description{
This is our hello world function!
}
\examples{
hello("Lindsay")


hello("Laura")
 hello("Laura","OWI")
 hello(last = "OWI", first = "Laura")

}
