% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(first, last = "Userbot")
}
\arguments{
\item{first}{character The first name that is said hello to.}

\item{last}{character The last name}
}
\description{
This is our hello world function!
}
\examples{
hello("Laura")
hello("Becky","OWI")
hello(last = "OWI", first = "Laura")
}
