% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{@title hello}
\usage{
hello(first, last = "userbot")
}
\description{
@description says hello world
 @param first character the first name that is said
 @param last character the last name that is said
 @export
}
\details{
@examples
 hello()
 hello("paul")
 hello("paul","R")
 hello(NULL)
}
