% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(first, last = "Userbot")
}
\arguments{
\item{first}{character The first name to say hello to.}

\item{last}{character the last name to say hello to.}
}
\description{
This is our hello world function!
}
\examples{
#put a basic example above the function
hello("Laura")
hello("Laura","OWI")
hello(last="OWI", first = "Laura")
}
