% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(user1, user2)
}
\arguments{
\item{user1}{character representing user 1}

\item{user2}{character representing user 2}
}
\description{
this is our hello world function
}
\examples{
hello("tom","sally")
hello("sally","tom")
}
