% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World Program}
\usage{
hello(a, b)
}
\arguments{
\item{a}{first variable for sum}

\item{b}{second variable for sum}
}
\value{
sum sum of input variables
}
\description{
This program is just to create a test package
}
\details{
this is a dummy function for hello world which generates sum of two variables
}
\author{
Trina Dey
}
