% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_dv.R
\name{summary_dv}
\alias{summary_dv}
\title{summary_dv}
\usage{
summary_dv(site)
}
\arguments{
\item{site}{character USGS site id}
}
\description{
Calculate min max of day of year
}
\examples{
sum_table <- summary_dv("08279500")

}
