% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_dv.R
\name{summary_dv}
\alias{summary_dv}
\title{summary_dv}
\usage{
summary_dv(site)
}
\arguments{
\item{site}{character USGS site ID}
}
\description{
Get max and min of discharge for day of the year.
}
\examples{
site <- "08279500"
sum_dv <- summary_dv(site)
}
